Ensure that all feature groups have at least two elements (#309)
* Ensure that all feature groups have at least two elements

This was harder to do than I thought. The problem is that the list of
features has to match what is in the CSV file. But in our case, that is
explicitly not what we want to do.

A slightly confusing aspect of this solution now is that
`disc_feature_groups` and `discrete_features` are a bit out of sync.
`disdisc_feature_groups` has some features that are not listed in
`discrete_features`. I don't think it's too bad though because code that
makes use of `discrete_features` is very low level code anyway and the
programmer who writes it can be expected to know about this subtlety.

* Allow single feature in feature group

* Turn off dummy generation by default and order on demand