Make datasets more like data structures than objects (#313)
Inspired by "Clean Code" and in particular the advice "Prefer data
structures", I turned removed all inheritance from the Dataset object.
This wasn't difficult, because the only thing the child classes were
overwriting was the `__init__()`. So, it was essentially used like a data
structure anyway. But now it is explicit. Instead of classes like Adult,
we now have functions like `adult()` which return a Dataset object
containing all the relevant information.

This change also removed some long standing weirdnesses in the Dataset
class (like having to initialize everything with empty lists).

For backwards compatibility (a thing we now kind of have to care about
since so much of our code is using EthicML), I added functions like
`Adult()` which for all intents and purposes act like the original class.
I only did that for our most used datasets though.

To keep the diff small, I kept a lot of code aritificially at the indentation
it had before.

Another nice thing about this change is that the factory functions can
return more than one thing. This made the code for the image datasets
much easier; which have to return the directory where the images are
stored somehow. Previously I solved this by subclassing Dataset as
ImageDataset. This is now not required anymore.