Turn the Results class into a NewType(DataFrame)
It was getting annoying to always do `results.data` to access the
underlying dataframe. Now that Results is just a NewType wrapper around
DataFrame, you don't have to do this anymore. The price we pay for this
is that all the class methods have to be turned into normal functions
which is a bit ugly, gut the "enduser" will likely never deal with them.
(The only place in the code where this comes up is in evaluate_models).