Write tests to simulate that no pytorch is available
This turned out to be a bit tricky. It seems the best way to do this is
to have a central point that checks if torch can be imported and which
then sets a global variable. Then, in the tests we can manipulate that
variable to make the code think that torch is not installed.