Remove `__init__` from algorithm base classes
The goal is to be able to use dataclasses for the algorithm classes.
Instead of Abstract Base Classes the algorithm base classes are now
Protocols. The reason for that is that protocols allow "abstract"
attributes (that is attributes that the child class must supply). I
already converted some of the algorithms to dataclasses.